requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
black==25.9.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
crewai==0.1.32
dataclasses-json==0.6.7
deepseek-sdk==0.1.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.121.0
flake8==7.3.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
loguru==0.7.3
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.10.1
numpy==1.26.4
openai==1.109.1
packaging==23.2
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.50.0
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
